Refactor: change namespace naming
update project name
