+++ QualificationsRegister.Tests/FunctionTests/QualificationsFunctionTests.cs
@@ -1,15 +1,14 @@
 using Microsoft.Azure.Functions.Worker;
-using Microsoft.Azure.Functions.Worker.Http;
-using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Moq;
-using QualificationsRegister.Functions;
-using QualificationsRegister.Tests.Mocks;
-using QualificationsRegister.UseCase.Interfaces;
-namespace QualificationsRegister.Tests.Functions
+using Ofqual.Common.RegisterAPI.Functions;
+using Ofqual.Common.RegisterAPI.Tests.Mocks;
+using Ofqual.Common.RegisterAPI.UseCase.Interfaces;
+
+namespace Ofqual.Common.RegisterAPI.Tests.Functions
 {
     [TestFixture]
-    public class Function1FunctionTests
+    public class QualificationsFunctionTests
     {
         private Mock<FunctionContext> _functionContext;
         private Mock<IGetOrganisationsSearchUseCase> _searchUseCaseMock;
@@ -27,9 +26,9 @@ namespace QualificationsRegister.Tests.Functions
         [Test]
         public async Task FunctionReturnsOkResponse()
         {
-            var httpFunc = new Function1(new NullLoggerFactory(), _searchUseCaseMock.Object, _byNumberUseCaseMock.Object);
+            var httpFunc = new Qualifications(new NullLoggerFactory(), _searchUseCaseMock.Object, _byNumberUseCaseMock.Object);
             MockHttpRequestData requestData = new MockHttpRequestData(_functionContext.Object);
-            var res = await httpFunc.Run(requestData);
+            var res = await httpFunc.ListQualifications(requestData);
             Console.WriteLine(res.StatusCode);
             Assert.That(res.StatusCode, Is.EqualTo(System.Net.HttpStatusCode.OK));
         }
