--- QualificationsRegister.Tests/UnitTest1.cs
@@ -1,16 +1,0 @@
-namespace QualificationsRegister.Tests
-{
-    public class Tests
-    {
-        [SetUp]
-        public void Setup()
-        {
-        }
-
-        [Test]
-        public void Test1()
-        {
-            Assert.Pass();
-        }
-    }
-}
